Move utilities to 'utils' module
