+++ paradigm/definitions/__init__.py
@@ -1,3 +1,2 @@
 from .base import (is_supported,
                    stdlib_modules_names)
-from .utils import to_contents
